refactor(acp): split transport_acp.go; ack configChangeFailed in Rust dispatch
Closes audit W7 + N2.

W7: transport_acp.go was 303 lines after the capability-negotiation
addition, past the project's 300-line "never acceptable" threshold.
Pulled two cohesive slices into sibling files:

  transport_acp_session.go     NewSession + SendPrompt +
                               RespondToPermission + CancelSession
                               (standard ACP session methods)
  transport_acp_control.go     SendControlRequest +
                               parseControlRequestCapability
                               (AgentsMesh extension already extracted)

transport_acp.go shrinks to 129 lines (struct + factory + Initialize
+ Handshake + ReadLoop + dispatchMessage + Close — the runtime
plumbing common to every ACP method).

N2: Rust acp_dispatch.rs treated configChangeFailed as unknown,
emitting a `tracing::warn!` per event. The web side already routes
the event to the activity-stream log surface; the Rust path needs
no state mutation, just an explicit case so warn noise stops.
